requirements.txt
requests
python-dotenv
flask
spotipy
openai 
beautifulsoup4 
supabase 
numpy